refactor(companion): decompose CompanionBase into domain mixin modules
companion_base.py (2,300+ lines) is now the assembly point; the
implementation lives in focused modules:
- base_support.py: ResponseWaiter, _SeenCache, adv_type_to_flags,
  _fmt_path, PUSH_CALLBACK_KEYS (re-exported from companion_base for
  backwards compatibility)
- base_contacts.py: contact/channel/path CRUD and advert-store logic
- base_config.py: device config, signing, flood scope, stats, custom vars
- base_callbacks.py: push-callback registration and response routing
- base_send.py: unified TX/send operations
- base_events.py: mesh-event RX handling and dedup

CompanionBase keeps its name, module path, public/protected surface, and
the _save_prefs/_load_prefs hooks, so CompanionRadio/CompanionBridge and
downstream subclasses are unaffected.
