+++ src/openhop_core/companion/base_config.py
@@ -1,0 +1,331 @@
+"""Device configuration, signing, flood scope, and stats of CompanionBase."""
+
+from __future__ import annotations
+
+import copy
+import logging
+import time
+from typing import Any, Optional
+
+from ..protocol import Packet
+from ..protocol.constants import ROUTE_TYPE_FLOOD, ROUTE_TYPE_TRANSPORT_FLOOD
+from ..protocol.transport_keys import calc_transport_code, get_auto_key_for
+from .constants import MAX_SIGN_DATA_SIZE, STATS_TYPE_CORE, STATS_TYPE_PACKETS, STATS_TYPE_RADIO
+from .models import NodePrefs
+
+logger = logging.getLogger("CompanionBase")
+
+
+class _DeviceConfigMixin:
+    """Part of :class:`CompanionBase` (see companion_base.py)."""
+
+    # -------------------------------------------------------------------------
+    # Device Configuration
+    # -------------------------------------------------------------------------
+
+    def set_advert_name(self, name: str) -> None:
+        """Set the node's advertised name (max 31 chars)."""
+        self.prefs.node_name = name[:31]
+        self._save_prefs()
+        self._sync_our_node_name_to_handlers()
+
+    def _get_group_text_handler(self) -> Optional[Any]:
+        """Return the group text handler for name sync, or None. Override in Radio/Bridge."""
+        return None
+
+    def _sync_our_node_name_to_handlers(self) -> None:
+        """Sync node name to group text handler for echo detection."""
+        handler = self._get_group_text_handler()
+        if handler is not None:
+            handler.set_our_node_name(self.prefs.node_name)
+
+    def set_advert_latlon(self, lat: float, lon: float) -> None:
+        """Set the GPS coordinates included in advertisements."""
+        if not (-90.0 <= lat <= 90.0):
+            raise ValueError(f"Latitude out of range: {lat}")
+        if not (-180.0 <= lon <= 180.0):
+            raise ValueError(f"Longitude out of range: {lon}")
+        self.prefs.latitude = lat
+        self.prefs.longitude = lon
+        self._save_prefs()
+
+    def set_radio_params(self, freq_hz: int, bw_hz: int, sf: int, cr: int) -> bool:
+        """Set radio parameters (frequency, bandwidth, SF, CR)."""
+        if not (5 <= sf <= 12):
+            raise ValueError(f"Spreading factor out of range: {sf}")
+        if not (5 <= cr <= 8):
+            raise ValueError(f"Coding rate out of range: {cr}")
+        self.prefs.frequency_hz = freq_hz
+        self.prefs.bandwidth_hz = bw_hz
+        self.prefs.spreading_factor = sf
+        self.prefs.coding_rate = cr
+        self._save_prefs()
+        return True
+
+    def set_tx_power(self, power_dbm: int) -> bool:
+        """Set the transmit power in dBm."""
+        self.prefs.tx_power_dbm = power_dbm
+        self._save_prefs()
+        return True
+
+    def set_tuning_params(self, rx_delay: float, airtime_factor: float) -> None:
+        """Set RX delay and airtime factor tuning parameters."""
+        self.prefs.rx_delay_base = rx_delay
+        self.prefs.airtime_factor = airtime_factor
+        self._save_prefs()
+
+    def get_tuning_params(self) -> tuple[float, float]:
+        """Return the current (rx_delay, airtime_factor) tuning parameters."""
+        return (self.prefs.rx_delay_base, self.prefs.airtime_factor)
+
+    def get_radio_params(self) -> dict:
+        """Return current radio configuration (frequency, bandwidth, SF, CR, TX power, tuning).
+
+        Use this to fetch the radio configuration details. Keys match the arguments
+        to set_radio_params/set_tx_power/set_tuning_params: frequency_hz, bandwidth_hz,
+        spreading_factor, coding_rate, tx_power_dbm, rx_delay_base, airtime_factor.
+        """
+        return {
+            "frequency_hz": self.prefs.frequency_hz,
+            "bandwidth_hz": self.prefs.bandwidth_hz,
+            "spreading_factor": self.prefs.spreading_factor,
+            "coding_rate": self.prefs.coding_rate,
+            "tx_power_dbm": self.prefs.tx_power_dbm,
+            "rx_delay_base": self.prefs.rx_delay_base,
+            "airtime_factor": self.prefs.airtime_factor,
+        }
+
+    def get_time(self) -> int:
+        """Return the current device time as a Unix timestamp."""
+        return int(time.time() + self._time_offset)
+
+    def set_time(self, secs: int) -> bool:
+        """Set the device time.  Returns False if *secs* is in the past."""
+        current = self.get_time()
+        if secs < current:
+            return False
+        self._time_offset = secs - time.time()
+        return True
+
+    def set_other_params(
+        self,
+        manual_add: int,
+        telemetry_modes: int,
+        advert_loc_policy: int,
+        multi_acks: int,
+    ) -> None:
+        """Set additional node parameters (manual add, telemetry, location, multi-acks)."""
+        self.prefs.manual_add_contacts = manual_add
+        self.prefs.telemetry_mode_base = telemetry_modes & 0x03
+        self.prefs.telemetry_mode_location = (telemetry_modes >> 2) & 0x03
+        self.prefs.telemetry_mode_environment = (telemetry_modes >> 4) & 0x03
+        self.prefs.advert_loc_policy = advert_loc_policy
+        self.prefs.multi_acks = multi_acks
+        self._save_prefs()
+
+    def set_path_hash_mode(self, mode: int) -> None:
+        """Set path hash encoding mode (0=1-byte, 1=2-byte, 2=3-byte hashes)."""
+        self.prefs.path_hash_mode = mode
+        self._save_prefs()
+
+    def get_self_info(self) -> NodePrefs:
+        """Return a copy of the current node preferences."""
+        return copy.copy(self.prefs)
+
+    def get_public_key(self) -> bytes:
+        """Return this node's 32-byte Ed25519 public key."""
+        return self._identity.get_public_key()
+
+    # -------------------------------------------------------------------------
+    # Signing Pipeline
+    # -------------------------------------------------------------------------
+
+    def sign_start(self) -> int:
+        """Begin a signing session; returns the maximum sign buffer size."""
+        self._sign_buffer = bytearray()
+        return MAX_SIGN_DATA_SIZE
+
+    def sign_data(self, data: bytes) -> bool:
+        """Append data to the signing buffer."""
+        if self._sign_buffer is None:
+            logger.warning("sign_data called without sign_start")
+            return False
+        if len(self._sign_buffer) + len(data) > MAX_SIGN_DATA_SIZE:
+            logger.warning("Sign data would overflow buffer")
+            return False
+        self._sign_buffer.extend(data)
+        return True
+
+    def sign_finish(self) -> Optional[bytes]:
+        if self._sign_buffer is None:
+            logger.warning("sign_finish called without sign_start")
+            return None
+        try:
+            return self._identity.sign(bytes(self._sign_buffer))
+        except Exception as e:
+            logger.error("Signing error: %s", e)
+            return None
+        finally:
+            self._sign_buffer = None
+
+    # -------------------------------------------------------------------------
+    # Key Management
+    # -------------------------------------------------------------------------
+
+    def export_private_key(self) -> bytes:
+        """Return the raw signing key bytes for backup/export."""
+        return self._identity.get_signing_key_bytes()
+
+    # -------------------------------------------------------------------------
+    # Flood Scope
+    # -------------------------------------------------------------------------
+
+    def set_flood_scope(self, transport_key: Optional[bytes] = None) -> None:
+        """Set or clear the flood transport key for scoped flooding.
+
+        Also cancels any pending explicit-unscoped request (firmware sets
+        ``send_unscoped = false`` whenever a scope override is set or reset).
+        """
+        if transport_key and len(transport_key) >= 16:
+            self._flood_transport_key = transport_key[:16]
+        else:
+            self._flood_transport_key = None
+        self._flood_unscoped = False
+
+    def set_flood_unscoped(self) -> None:
+        """Force the next flood to be unscoped, bypassing the default scope.
+
+        Mirrors firmware CMD_SET_FLOOD_SCOPE_KEY mode 1 (FW PR #2492): a one-shot
+        flag consumed by the next flood-routed packet in _apply_flood_scope.
+        """
+        self._flood_unscoped = True
+
+    def set_default_flood_scope(
+        self,
+        scope_name: Optional[str],
+        transport_key: Optional[bytes],
+    ) -> bool:
+        """Persist default flood scope (v1.15 companion protocol semantics)."""
+        if not scope_name or not transport_key or len(transport_key) < 16:
+            self.prefs.default_scope_name = ""
+            self.prefs.default_scope_key = b""
+            self._save_prefs()
+            return True
+        normalized = scope_name[:30].strip()
+        if not normalized:
+            return False
+        self.prefs.default_scope_name = normalized
+        self.prefs.default_scope_key = bytes(transport_key[:16])
+        self._save_prefs()
+        return True
+
+    def get_default_flood_scope(self) -> Optional[tuple[str, bytes]]:
+        """Return (name, key) for persisted default scope, or None if unset."""
+        name = (getattr(self.prefs, "default_scope_name", "") or "").strip()
+        key = getattr(self.prefs, "default_scope_key", b"") or b""
+        key = bytes(key[:16]).ljust(16, b"\x00") if key else b""
+        if not name or key == b"\x00" * 16:
+            return None
+        return (name, key)
+
+    def set_flood_region(self, region_name: Optional[str] = None) -> None:
+        """Set flood scope from a region name (e.g., ``'#usa'``) or clear it.
+
+        Derives the 16-byte transport key automatically via SHA-256 of the
+        region name.  A leading ``#`` is added if not already present.
+        Pass ``None`` to clear the scope (flood to all).
+        """
+        if region_name:
+            if not region_name.startswith("#"):
+                region_name = f"#{region_name}"
+            self._flood_transport_key = get_auto_key_for(region_name)
+        else:
+            self._flood_transport_key = None
+
+    def _resolve_flood_transport_key(self) -> Optional[bytes]:
+        """Resolve effective flood key: transient key first, then persisted default."""
+        if self._flood_transport_key is not None:
+            return self._flood_transport_key
+        default_scope = self.get_default_flood_scope()
+        if default_scope is None:
+            return None
+        return default_scope[1]
+
+    def _apply_flood_scope(self, pkt: Packet) -> None:
+        """Apply flood scope transport codes to a packet in-place.
+
+        If ``_flood_transport_key`` is set and the packet uses flood routing,
+        calculates the transport code, attaches it to the packet, and changes
+        the route type to ``ROUTE_TYPE_TRANSPORT_FLOOD``.
+
+        Matches firmware ``sendFloodScoped()`` in ``BaseChatMesh.cpp``.
+        """
+        route_type = pkt.get_route_type()
+        if route_type != ROUTE_TYPE_FLOOD:
+            return  # only scope flood packets, not direct
+        if self._flood_unscoped:
+            # App explicitly requested unscoped (FW #2492): leave as plain flood,
+            # ignoring any default scope. One-shot — consumed here, as firmware
+            # resets send_unscoped after each flood.
+            self._flood_unscoped = False
+            return
+        effective_key = self._resolve_flood_transport_key()
+        if effective_key is None:
+            return
+        code = calc_transport_code(effective_key, pkt)
+        pkt.transport_codes[0] = code
+        pkt.transport_codes[1] = 0  # reserved for home region (firmware TODO)
+        # Switch route type from FLOOD -> TRANSPORT_FLOOD
+        pkt.header = (pkt.header & ~0x03) | ROUTE_TYPE_TRANSPORT_FLOOD
+
+    def _apply_path_hash_mode(self, pkt: Packet) -> None:
+        """Encode the device's path_hash_mode in originated packets.
+
+        When a packet has 0 hops (freshly originated), sets bits 6-7 of
+        ``path_len`` to encode the hash size from ``prefs.path_hash_mode``.
+        Packets with existing hops (stored contact paths) are untouched.
+        Trace packets are excluded because the repeater's trace handler uses
+        ``path``/``path_len`` to store SNR values, not routing hashes.
+        Sets ``_path_hash_mode_applied`` so the dispatcher does not overwrite.
+        """
+        pkt.apply_path_hash_mode(self.prefs.path_hash_mode, mark_applied=True)
+
+    # -------------------------------------------------------------------------
+    # Statistics (subclasses may override _get_radio_stats for STATS_TYPE_RADIO)
+    # -------------------------------------------------------------------------
+
+    def get_stats(self, stats_type: int = STATS_TYPE_PACKETS) -> dict:
+        """Return statistics of the requested type (core, radio, or packets)."""
+        if stats_type == STATS_TYPE_CORE:
+            return {
+                "uptime_secs": self.stats.get_uptime_secs(),
+                "queue_len": self.message_queue.count,
+                "contacts_count": self.contacts.get_count(),
+                "channels_count": self.channels.get_count(),
+            }
+        elif stats_type == STATS_TYPE_RADIO:
+            return self._get_radio_stats()
+        return self.stats.get_totals()
+
+    def _get_radio_stats(self) -> dict:
+        """Override in CompanionRadio for hardware RSSI/SNR. Default: prefs only."""
+        return {
+            "frequency_hz": self.prefs.frequency_hz,
+            "bandwidth_hz": self.prefs.bandwidth_hz,
+            "spreading_factor": self.prefs.spreading_factor,
+            "coding_rate": self.prefs.coding_rate,
+            "tx_power_dbm": self.prefs.tx_power_dbm,
+        }
+
+    # -------------------------------------------------------------------------
+    # Custom Variables
+    # -------------------------------------------------------------------------
+
+    def get_custom_vars(self) -> dict[str, str]:
+        """Return a copy of all custom variables."""
+        return dict(self._custom_vars)
+
+    def set_custom_var(self, name: str, value: str) -> bool:
+        """Set a custom variable by name."""
+        self._custom_vars[name] = value
+        return True
