+++ src/openhop_core/companion/contact_store.py
@@ -277,9 +277,9 @@ class ContactStore:
                 name=rec.get("name", ""),
                 adv_type=rec.get("adv_type", 0),
                 flags=rec.get("flags", 0),
-                out_path_len=-1
-                if rec.get("out_path_len", -1) in (-1, 255)
-                else rec.get("out_path_len", -1),
+                out_path_len=(
+                    -1 if rec.get("out_path_len", -1) in (-1, 255) else rec.get("out_path_len", -1)
+                ),
                 out_path=out_path,
                 last_advert_timestamp=rec.get("last_advert_timestamp", 0),
                 lastmod=rec.get("lastmod", 0),
@@ -311,9 +311,9 @@ class ContactStore:
                     "gps_lat": c.gps_lat,
                     "gps_lon": c.gps_lon,
                     "sync_since": c.sync_since,
-                    "last_advert_packet": c.last_advert_packet.hex()
-                    if c.last_advert_packet
-                    else "",
+                    "last_advert_packet": (
+                        c.last_advert_packet.hex() if c.last_advert_packet else ""
+                    ),
                 }
             )
         return result
